fix: improved API, focus mgmt, widget layout
The onChanged() callback now receives the updated list of values. This
allows the widget to take a constant list of initial values as the state no
longer modifies the list in-place. Similarly, the findSuggestions() callback
can return an immutable list as well. No more sharing state between the
widget/state and the application.

Added some space between the text field and the input chips, defaulting to 8
pixels, and controlled by the new property interiorSpacing.

Fixed the focus management such that for both selecting an existing value or
creating a new one will return the focus back to the text field.