feat: optional placeholder item to avoid shifting
Without the optional placeholder item, the widgets below the FlutterTagging
widget will shift down once chips are added. Similarly, if the chips are
removed then the widgets below will shift up. To prevent this distracting
behavior, an optional placeholderItem may be provided that will create an
invisible chip that holds the screen space in reserve.